+++ src/views/AgentEdit.vue
@@ -46,11 +46,6 @@
             class="pt-2"
             style="display: flex; flex-direction: row; align-items: center"
           >
-            <agent-script-import-dialog
-              v-model="scriptImportDialog"
-              :loading="scriptImportLoading"
-              @submit="doScriptImport"
-            />
             <agent-upload-dialog
               v-model="uploadDialog"
               :language="agent.language"
@@ -83,11 +78,6 @@
               icon="fa-download"
               text="Download"
               @click="downloadDialog = true"
-            />
-            <tooltip-button
-              icon="fa-file-import"
-              text="Import Script"
-              @click="scriptImportDialog = true"
             />
             <tooltip-button
               v-if="!hideSideBar"
@@ -252,7 +242,6 @@ import AgentExecuteModule from "@/components/agents/AgentExecuteModule.vue";
 import AgentFileBrowser from "@/components/agents/AgentFileBrowser.vue";
 import AgentTerminal from "@/components/agents/AgentTerminal.vue";
 import AgentUploadDialog from "@/components/agents/AgentUploadDialog.vue";
-import AgentScriptImportDialog from "@/components/agents/AgentScriptImportDialog.vue";
 import AgentDownloadDialog from "@/components/agents/AgentDownloadDialog.vue";
 import TooltipButton from "@/components/TooltipButton.vue";
 import TooltipButtonToggle from "@/components/TooltipButtonToggle.vue";
@@ -274,7 +263,6 @@ export default {
     AgentTasksList,
     AgentTerminal,
     AgentUploadDialog,
-    AgentScriptImportDialog,
     AgentDownloadDialog,
     TooltipButton,
     TooltipButtonToggle,
@@ -431,21 +419,6 @@ export default {
         .catch(() => {
           this.errorState = true;
         });
-    },
-    async doScriptImport({ file }) {
-      if (this.scriptImportLoading || file == null) return;
-
-      this.scriptImportLoading = true;
-      try {
-        await agentTaskApi.scriptImport(this.agent.session_id, file);
-        this.$snack.success(
-          `Tasked agent ${this.agent.name} to import script ${file.filename}`,
-        );
-      } catch (err) {
-        this.$snack.error(`Error: ${err}`);
-      }
-      this.scriptImportLoading = false;
-      this.scriptImportDialog = false;
     },
     async killAgent() {
       if (